Move inRange to ChartUtils file
I want to reuse this function from Redux selector later so I need it to be independent from generateCategoricalChart
